seperated manager view from user creation , stricter roles
When you were editing or creating a new employee a pane appeared that was actually part of the managerView. Now they are separate. I also replaced the textfield for editing the role to a combobox. Now you only have a couple options to chose from. With this comes a Role class that is an enum that holds all of the roles (Its in utilities).