Titlebar stuttering when you pulled the left side, Fix
When first implemented the title bar, it worked just fine on my laptop. But when I tried it on my  pc, it started to stutter when I pulled the left side of the screen when trying to enlarge it. I changed the way the code works so hopefully this fixes it.